Migrate to tauri beta
