+++ package.json
@@ -17,6 +17,8 @@
     "@codemirror/lang-markdown": "^6.2.4",
     "@codemirror/state": "^6.4.0",
     "@codemirror/view": "^6.24.0",
+    "@tauri-apps/plugin-dialog": "2.0.0-beta.2",
+    "@tauri-apps/plugin-fs": "2.0.0-beta.2",
     "@uiw/codemirror-theme-github": "^4.21.21",
     "@uiw/codemirror-themes": "^4.21.21",
     "@uiw/react-codemirror": "^4.21.21",
@@ -29,8 +31,8 @@
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.10",
     "@tailwindcss/vite": "4.0.0-alpha.3",
-    "@tauri-apps/api": "^1.5.3",
-    "@tauri-apps/cli": "^1.5.10",
+    "@tauri-apps/api": "2.0.0-beta.5",
+    "@tauri-apps/cli": "2.0.0-beta.9",
     "@types/node": "20.11.17",
     "@types/react": "^18.2.55",
     "@types/react-dom": "^18.2.19",
