+++ src-tauri/Cargo.toml
@@ -12,16 +12,22 @@ rust-version = "1.60"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
-tauri-build = { version = "1.5.1", features = [] }
+tauri-build = { version = "2.0.0-beta", features = [] }
 
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
-tauri = { version = "1.6.0", features = [ "path-all", "fs-all", "dialog-all"] }
+tauri = { version = "2.0.0-beta", features = [] }
 comrak = "0.21.0"
+tauri-plugin-fs = "2.0.0-beta.2"
+tauri-plugin-dialog = "2.0.0-beta.2"
 
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem and the built-in dev server is disabled.
 # If you use cargo directly instead of tauri's cli you can use this feature flag to switch between tauri's `dev` and `build` modes.
 # DO NOT REMOVE!!
 custom-protocol = [ "tauri/custom-protocol" ]
+
+[lib]
+name = "markdone_app_lib"
+crate-type = ["staticlib", "cdylib", "lib"]
